Display multiple recipes on home screen and allow navigation to each one
